Clean core-test and fix linkage errors on older gcc
