+++ test/CMakeLists.txt
@@ -85,9 +85,9 @@ function(add_fmt_test name)
 endfunction()
 
 add_fmt_test(assert-test)
-add_fmt_test(core-test mock-allocator.h)
+add_fmt_test(core-test)
 add_fmt_test(gtest-extra-test)
-add_fmt_test(format-test)
+add_fmt_test(format-test mock-allocator.h)
 add_fmt_test(format-impl-test)
 add_fmt_test(ostream-test)
 add_fmt_test(printf-test)
